pyproject.toml
# Please don't edit: generated by nmk-python plugin

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nmk-base"
dynamic = ["version"]
requires-python = ">=3.9"
classifiers = ["Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13"]
dependencies = ["Jinja2", "nmk>=1.1.0", "buildenv", "tomlkit"]
description = "Base plugin for nmk build system"

[[project.authors]]
name = "The dynod project"

[[project.maintainers]]
name = "The dynod project"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
text = "MIT License"

[project.urls]
Homepage = "https://github.com/dynod/nmk-base"
Repository = "https://github.com/dynod/nmk-base"
Issues = "https://github.com/dynod/nmk/issues?q=is%3Aopen+is%3Aissue+label%3Aplugin%3Abase"

[project.entry-points.buildenv_init]
nmk = "nmk_base._buildenv.legacy:BuildEnvInit"

[project.entry-points.buildenv_extension]
nmk = "nmk_base._buildenv.extension:NmkBaseBuildEnvExtension"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*"]

[tool.setuptools.dynamic.version]
file = "out/.pythonversion"

[tool.distutils.bdist_wheel]
plat-name = "any"

[tool.pyright]
include = ["src"]
typeCheckingMode = "strict"
reportMissingTypeStubs = false

[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E203", "E501"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
junit_family = "xunit2"
filterwarnings = ["ignore::DeprecationWarning"]
addopts = ["--capture=sys", "--junit-xml=out/tests/report.xml", "--cov=src", "--cov-report=html:out/coverage-report", "--cov-report=xml:out/coverage.xml", "--cov-report=term", "--cov-config=pyproject.toml"]

[tool.coverage.run]
data_file = "out/.coverage"
concurrency = ["thread"]
disable_warnings = ["no-data-collected"]
branch = true
omit = ["src/tests/*"]
